:paperclip:

+++ examples/audio/spatial_audio_2d.rs
@@ -88,7 +88,7 @@ fn update_positions(
 ) {
     if let Some(sink) = audio_sinks.get(&music_controller.0) {
         let mut emitter_transform = emitter.single_mut();
-        emitter_transform.translation.x = (time.elapsed_seconds()).sin() as f32 * 5.0;
+        emitter_transform.translation.x = time.elapsed_seconds().sin() * 5.0;
         sink.set_emitter_position(emitter_transform.translation);
     }
 }
